% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_plots.R
\name{biplot_raw_source}
\alias{biplot_raw_source}
\title{Raw Source Biplot}
\usage{
biplot_raw_source(
  data = refvals,
  title = "IsoSpaceFoodSourceDist",
  united = FALSE,
  shrink_leg = FALSE
)
}
\arguments{
\item{data}{Reference Values}

\item{title}{title of plot}

\item{united}{Unite plot and legend?  Default is FALSE}

\item{shrink_leg}{Shrink legend?  Default is FALSE}
}
\value{
Plot
}
\description{
Raw Source Biplot
}
\examples{
# biplot_raw_source()
}
