% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_plots.R
\name{biplot_raw_sourcon}
\alias{biplot_raw_sourcon}
\title{Raw Source and Consumer Biplot}
\usage{
biplot_raw_sourcon(
  refdata = refvals,
  mixdata,
  title = "IsoSpaceRawSourceandConsumer",
  Sites = 1,
  timer = "Stat",
  labelID = TRUE
)
}
\arguments{
\item{refdata}{Source data}

\item{mixdata}{Cosumer data}

\item{title}{Title}

\item{Sites}{Number of sites}

\item{timer}{Time stratification variable}

\item{labelID}{Label ID (TRUE) or not (FALSE)?  Default is TRUE}
}
\value{
Plot
}
\description{
Raw Source and Consumer Biplot
}
\examples{
#Add later
}
