% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_isos.R
\name{comp_isos}
\alias{comp_isos}
\title{Compare Isos}
\usage{
comp_isos(data, comp, by, re = NULL, mixed = FALSE)
}
\arguments{
\item{data}{Data frame}

\item{comp}{Outcome variable}

\item{by}{Main comparison variable}

\item{re}{Random intercept}

\item{mixed}{Set to TRUE to use nlme::lme for statistical test (default is
FALSE)}
}
\value{
Statistical Test results
}
\description{
Compare Isos
}
\examples{
#Add later
}
