% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{discrimination}
\alias{discrimination}
\title{TDF Sider Discrimination}
\format{
\subsection{\code{discrimination}}{

A data frame with 4 rows and 5 columns:
\describe{
\item{Group}{#}
\item{Meand15N}{#}
\item{SDd15N}{#}
\item{Meand13C}{#}
\item{SDd13C}{#}
}
}
}
\usage{
discrimination
}
\description{
TDF Sider Discrimination
}
\keyword{datasets}
