% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_plots.R
\name{main_box}
\alias{main_box}
\title{Main Boxplot Function}
\usage{
main_box(mixdata)
}
\arguments{
\item{mixdata}{Consumer data}
}
\value{
Plot
}
\description{
Main Boxplot Function
}
\examples{
# Add later
}
