% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refvals}
\alias{refvals}
\title{Reference Values}
\format{
\subsection{\code{refvals}}{

A data frame with 245 rows and 5 columns:
\describe{
\item{Group}{#}
\item{Organism}{#}
\item{d15N}{#}
\item{d13C}{#}
\item{Paper Source}{#}
}
}
}
\usage{
refvals
}
\description{
Reference Values
}
\keyword{datasets}
