% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_plots.R
\name{source_biplot}
\alias{source_biplot}
\title{Source Biplot}
\usage{
source_biplot(
  data = refvals,
  Group = Group,
  var1 = d13C,
  var2 = d15N,
  title = "IsoSpaceSourceSummary"
)
}
\arguments{
\item{data}{Source data}

\item{Group}{Group variable}

\item{var1}{var 1}

\item{var2}{var 2}

\item{title}{Title}
}
\value{
Plot
}
\description{
Source Biplot
}
\examples{
# Add later
}
