% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_plots.R
\name{sourcecon_biplot}
\alias{sourcecon_biplot}
\title{Source and Consumer Biplot}
\usage{
sourcecon_biplot(refdata = refvals, mixdata, Sites = 1, title = "")
}
\arguments{
\item{refdata}{Source data}

\item{mixdata}{Consumer data}

\item{Sites}{Number of sites}

\item{title}{Title}
}
\value{
A plot
}
\description{
Source and Consumer Biplot
}
\examples{
#Add later
}
