% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tab2}
\alias{tab2}
\title{Tab 2: OM Full Comparison}
\format{
\subsection{\code{tab2}}{

A data frame with 98 rows and 6 columns:
\describe{
\item{Site}{#}
\item{UniqueID}{#}
\item{d15N}{#}
\item{d13C}{#}
\item{Period}{#}
\item{Stat}{#}
\item{Stat Period}{#}
}
}
}
\usage{
tab2
}
\description{
The Full comparison has the data broken up into peasants (0), elite (1), and
monks (2). This is going to compare these socioeconomic status to each other.
The time portion does not get utilized in this one.
}
\keyword{datasets}
