% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tab3}
\alias{tab3}
\title{Tab 3: OM Peasant/Time Comparison}
\format{
\subsection{\code{tab3}}{

A data frame with 55 rows and 6 columns:
\describe{
\item{Site}{#}
\item{UniqueID}{#}
\item{d15N}{#}
\item{d13C}{#}
\item{Period}{#}
\item{Stat}{#}
\item{Stat Period}{#}
}
}
}
\usage{
tab3
}
\description{
This comparison is looking only at peasants (0) and comparing them through time
which is signified by the Arm Position/Periods A, B, and C/D.  Three time periods
of actual time are listed as Pre 1300 CE, 1300 CE - 1375 CE, and 1375 CE to close
of the period (I will get an exact date on this).
}
\keyword{datasets}
