% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tab4}
\alias{tab4}
\title{Tab 4: OM Monk/Time Comparison}
\format{
\subsection{\code{tab4}}{

A data frame with 13 rows and 7 columns:
\describe{
\item{Site}{#}
\item{UniqueID}{#}
\item{d15N}{#}
\item{d13C}{#}
\item{Period}{#}
\item{Stat}{#}
\item{Stat Period}{#}
}
}
}
\usage{
tab4
}
\description{
Looking at only Monks (2) and comparing them over the three time periods
(A, B, and C/D)
}
\keyword{datasets}
