% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tab5}
\alias{tab5}
\title{Tab 5:OM Elite/Time Comparison}
\format{
\subsection{\code{tab5}}{

A data frame with 29 rows and 7 columns:
\describe{
\item{Site}{#}
\item{UniqueID}{#}
\item{d15N}{#}
\item{d13C}{#}
\item{Period}{#}
\item{Stat}{#}
\item{Stat Period}{#}
}
}
}
\usage{
tab5
}
\description{
Looking at only Elite (1) and comparing them over the three time periods
(A, B, and C/D).
}
\keyword{datasets}
