% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tab6}
\alias{tab6}
\title{Tab 6:SM Peasant/Time Comparison}
\format{
\subsection{\code{tab6}}{

A data frame with 59 rows and 7 columns:
\describe{
\item{Site}{#}
\item{UniqueID}{#}
\item{d15N}{#}
\item{d13C}{#}
\item{Period}{#}
\item{Stat}{#}
\item{Stat Period}{#}
}
}
}
\usage{
tab6
}
\description{
SM is St. Mikkel.  This is comparing Peasants (0) and comparing them over the
three time periods (A, B, and C/D).  There were no samples of monks or elite
in SM so no additional comparisons.
}
\keyword{datasets}
