% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tab7}
\alias{tab7}
\title{Tab 7:Ribe Peasant/Time Comparison}
\format{
\subsection{\code{tab7}}{

A data frame with 54 rows and 7 columns:
\describe{
\item{Site}{#}
\item{UniqueID}{#}
\item{d15N}{#}
\item{d13C}{#}
\item{Period}{#}
\item{Stat}{#}
\item{Stat Period}{#}
}
}
}
\usage{
tab7
}
\description{
Ribe is Ribe (not short for anything).  This is comparing Peasants (0) and
comparing them over the three time periods (A, B, and C/D).  There were no
samples of monks or elite in Ribe so no additional comparisons.
}
\keyword{datasets}
