% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tab8}
\alias{tab8}
\title{Tab 8:Tirup Peasant/Time Comparison}
\format{
\subsection{\code{tab8}}{

A data frame with 32 rows and 7 columns:
\describe{
\item{Site}{#}
\item{UniqueID}{#}
\item{d15N}{#}
\item{d13C}{#}
\item{Period}{#}
\item{Stat}{#}
\item{Stat Period}{#}
}
}
}
\usage{
tab8
}
\description{
Tirup is Tirup (again not short for anything). This is comparing Peasants (0)
and comparing them over the three time periods (A, B, and C/D).  There were no
samples of monks or elite in Tirup so no additional comparisons.
}
\keyword{datasets}
