% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tab9}
\alias{tab9}
\title{Tab 9: All Site Peasants/Time Comparison}
\format{
\subsection{\code{tab9}}{

A data frame with 200 rows and 7 columns:
\describe{
\item{Site}{#}
\item{UniqueID}{#}
\item{d15N}{#}
\item{d13C}{#}
\item{Period}{#}
\item{Stat}{#}
\item{Stat Period}{#}
}
}
}
\usage{
tab9
}
\description{
This takes the Peasants (0) from all 4 sites (OM, SM, Tirup, Ribe) and compares
them over the three time periods (A, B, and C/D).
}
\keyword{datasets}
