% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_plots.R
\name{biplot_raw_sourcon}
\alias{biplot_raw_sourcon}
\title{Raw Source and Consumer Biplot}
\usage{
biplot_raw_sourcon(refdata = refvals, mixdata, title = "", Sites = 1)
}
\arguments{
\item{refdata}{Source data}

\item{mixdata}{Cosumer data}

\item{title}{Title}

\item{Sites}{Number of sites}
}
\value{
Plot
}
\description{
Raw Source and Consumer Biplot
}
\examples{
#Add later
}
