% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_plots.R
\name{sourcecon_biplot}
\alias{sourcecon_biplot}
\title{Source and Consumer Biplot}
\usage{
sourcecon_biplot(
  refdata = refvals,
  mixdata,
  Sites = 1,
  title = "IsoSpaceSSRC",
  timer = "Stat",
  labelID = TRUE
)
}
\arguments{
\item{refdata}{Source data}

\item{mixdata}{Consumer data}

\item{Sites}{Number of sites}

\item{title}{Title}

\item{timer}{Time stratification variable}

\item{labelID}{Label ID (TRUE) or not (FALSE)?  Default is TRUE}
}
\value{
A plot
}
\description{
Source and Consumer Biplot
}
\examples{
#Add later
}
